WIP: Write didTapWeightButton func.

+++ BarbellMath/PlateCountViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class PlateCountViewController: UIViewController {
     
-    var totalWeight: Int = 0
+    var totalWeight: Double = 0
     //should this be nil instead of 0? How would using an optional value here change things? See temp app...? Mike says I could solve this with a "guard let" statement (if total weight is 0) in the plate math func instead. Look up guard statements.
     
     @IBOutlet weak var text45LbPlates: UILabel!
